parse_key do not break if empty or undefined dict
