+++ macapype/utils/misc.py
@@ -58,43 +58,17 @@ def merge_3_elem_to_list(elem1, elem2, elem3):
     return [elem1, elem2, elem3]
 
 
-def get_dict_from_json(json_file):
+#def get_dict_from_json(json_file):
 
-    import os
-    import json
+    #import os
+    #import json
 
-    print(json_file)
+    #print(json_file)
 
-    assert os.path.exists(json_file), "Error with file {}".format(json_file)
-    params = json.load(open(json_file))
+    #assert os.path.exists(json_file), "Error with file {}".format(json_file)
+    #params = json.load(open(json_file))
 
-    assert "macapype" in params.keys(),\
-        "Error, json should contain 'macapype' field"
+    #assert "macapype" in params.keys(),\
+        #"Error, json should contain 'macapype' field"
 
-    return params["macapype"]
-
-
-def parse_key(params, key):
-
-    from nipype.interfaces.base import isdefined
-
-    def _parse_key(params, cur_key):
-        if  cur_key in params.keys():
-            return params[cur_key]
-        else:
-            "Error, key {} was not found in {}".format(key, params.keys())
-            return {}
-
-    if isdefined(params):
-        if isinstance(key, tuple):
-            for cur_key in key:
-                params = _parse_key(params, cur_key)
-
-        else:
-            params = _parse_key(params, key)
-
-        return params
-
-    else:
-        return {}
-
+    #return params["macapype"]
